+++ sangyun/MySQL_API/app.js
@@ -3,6 +3,8 @@ const express = require('express');
 const dotenv = require('dotenv');
 dotenv.config();
 
+const {getBook, addBook} = require('./book.controller');
+
 const app = express();
 app.use(express.json());
 
@@ -32,53 +34,9 @@ dataSource.initialize().then(() => {
 });
 
 //
-app.get('/book', async (req, res) => {
-  const answer = await dataSource.query(`SELECT * FROM books`)
-                        .catch((err) => {
-                          console.log("query has failed: ", err.sqlMessage);
-                          res.status(500);
-                          return Promise.resolve([]);
-                        });
 
-  const msg = answer.map((book) => {
-    const { title, description, cover_image, created_at, updated_at } = book;
-    return {
-      title,
-      description,
-      cover_image,
-      created_at,
-      updated_at,
-    };
-  });
-
-  res.send(msg);
-})
-
-app.post("/book", async (req, res) => {
-  const {title, description, coverImage='covDft'} = req.body;
-  if (!title | !description) {
-    res.status(400).json({ message: "title or description is missing" });
-    return;
-  }
-
-  const answer = await dataSource.query(
-                                  `INSERT INTO books(
-                                    title,
-                                    description,
-                                    cover_image
-                                  ) VALUES (?, ?, ?);
-                                  `,
-                                  [title, description, coverImage]
-                                ).catch((err) => (Promise.resolve(undefined)))
-
-  console.dir(answer);
-
-  if (answer) {
-    res.status(201).json({ message: "successfully created" });
-  } else {
-    res.status(500).json({ message: "failed to create" });
-  }
-});
+app.get('/book', getBook(dataSource));
+app.post("/book", addBook(dataSource));
 
 // init
 const PORT = process.env.PORT || 5000;
